add redis context lib
